Remove initial cache POC.
