+++ packages/service-core/test/src/large_batch.test.ts
@@ -49,8 +49,8 @@ function defineBatchTests(factory: StorageFactory) {
       const checkpoint = await context.getCheckpoint({ timeout: 100_000 });
       const duration = Date.now() - start;
       const used = Math.round(process.memoryUsage().heapUsed / 1024 / 1024);
-      const checksum = await context.storage!.getChecksums(checkpoint, null, ['global[]']);
-      expect(checksum[0].count).toEqual(operation_count);
+      const checksum = await context.storage!.getChecksums(checkpoint, ['global[]']);
+      expect(checksum.get('global[]')!.count).toEqual(operation_count);
       const perSecond = Math.round((operation_count / duration) * 1000);
       console.log(`${operation_count} ops in ${duration}ms ${perSecond} ops/s. ${used}MB heap`);
     }),
@@ -100,8 +100,8 @@ function defineBatchTests(factory: StorageFactory) {
 
         const checkpoint = await context.getCheckpoint({ timeout: 100_000 });
         const duration = Date.now() - start;
-        const checksum = await context.storage!.getChecksums(checkpoint, null, ['global[]']);
-        expect(checksum[0].count).toEqual(operation_count);
+        const checksum = await context.storage!.getChecksums(checkpoint, ['global[]']);
+        expect(checksum.get('global[]')!.count).toEqual(operation_count);
         const perSecond = Math.round((operation_count / duration) * 1000);
         console.log(`${operation_count} ops in ${duration}ms ${perSecond} ops/s.`);
         printMemoryUsage();
@@ -156,8 +156,8 @@ function defineBatchTests(factory: StorageFactory) {
       const checkpoint = await context.getCheckpoint({ timeout: 50_000 });
       const duration = Date.now() - start;
       const used = Math.round(process.memoryUsage().heapUsed / 1024 / 1024);
-      const checksum = await context.storage!.getChecksums(checkpoint, null, ['global[]']);
-      expect(checksum[0].count).toEqual(operationCount);
+      const checksum = await context.storage!.getChecksums(checkpoint, ['global[]']);
+      expect(checksum.get('global[]')!.count).toEqual(operationCount);
       const perSecond = Math.round((operationCount / duration) * 1000);
       // This number depends on the test machine, so we keep the test significantly
       // lower than expected numbers.
@@ -173,8 +173,8 @@ function defineBatchTests(factory: StorageFactory) {
       const checkpoint2 = await context.getCheckpoint({ timeout: 20_000 });
       const truncateDuration = Date.now() - truncateStart;
 
-      const checksum2 = await context.storage!.getChecksums(checkpoint2, null, ['global[]']);
-      const truncateCount = checksum2[0].count - checksum[0].count;
+      const checksum2 = await context.storage!.getChecksums(checkpoint2, ['global[]']);
+      const truncateCount = checksum2.get('global[]')!.count - checksum.get('global[]')!.count;
       expect(truncateCount).toEqual(numTransactions * perTransaction);
       const truncatePerSecond = Math.round((truncateCount / truncateDuration) * 1000);
       console.log(`Truncated ${truncateCount} ops in ${truncateDuration}ms ${truncatePerSecond} ops/s. ${used}MB heap`);
